Fold BenchmarkOrcDecimalReader into BenchmarkBatchStreamReaders
```

Benchmark                                             Mode  Cnt  Score   Error  Units
BenchmarkBatchStreamReaders.readLongDecimalNoNull     avgt   60  0.829 ± 0.032   s/op
BenchmarkBatchStreamReaders.readLongDecimalWithNull   avgt   60  0.566 ± 0.015   s/op
BenchmarkBatchStreamReaders.readShortDecimalNoNull    avgt   60  0.544 ± 0.026   s/op
BenchmarkBatchStreamReaders.readShortDecimalWithNull  avgt   60  0.364 ± 0.006   s/op
```
